+++ tests/test_nludb.py
@@ -3,8 +3,7 @@ import os
 import random
 import string
 
-from nludb import NLUDB
-from nludb.types.embedding_index import EmbeddingModels
+from nludb import NLUDB, EmbeddingModels
 
 __author__ = "Edward Benson"
 __copyright__ = "Edward Benson"
@@ -15,19 +14,29 @@ def _random_name() -> str:
     id =''.join(random.choice(letters) for i in range(10))
     return "test_{}".format(id)
 
+def _nludb() -> NLUDB:
+    assert os.environ['NLUDB_KEY'] is not None
+    NLUDB_DOMAIN = None
+    if os.environ['NLUDB_DOMAIN'] is not None:
+        NLUDB_DOMAIN = os.environ['NLUDB_DOMAIN']
+    nludb = NLUDB(
+        api_key = os.environ['NLUDB_KEY'],
+        api_domain=NLUDB_DOMAIN
+    )
+    return nludb
+
 def test_connect():
     """Test basic connection"""
-    assert os.environ['NLUDB_KEY'] is not None
-    nludb = NLUDB(api_key = os.environ['NLUDB_KEY'])
+    nludb = _nludb()
 
 def test_index_create():
-    nludb = NLUDB(api_key = os.environ['NLUDB_KEY'])
+    nludb = _nludb()
     name = _random_name()
 
     # Should require name
     with pytest.raises(Exception):
         index = nludb.create_index(
-            model=EmbeddingModels.DEFAULT_QA
+            model=EmbeddingModels.QA
         )
 
     # Should require model
@@ -38,7 +47,7 @@ def test_index_create():
 
     index = nludb.create_index(
         name=name,
-        model=EmbeddingModels.DEFAULT_QA,
+        model=EmbeddingModels.QA,
         upsert=True
     )
     assert index is not None
@@ -47,17 +56,74 @@ def test_index_create():
     with pytest.raises(Exception, match=r".*already exists.*"):
         index = nludb.create_index(
             name="Test Index",
-            model=EmbeddingModels.DEFAULT_QA,
+            model=EmbeddingModels.QA,
             upsert=False
         )
 
+def test_basic_embeddings():
+    nludb = _nludb()
+    e1 = nludb.embed(["This is a test"], EmbeddingModels.QA)
+    e1b = nludb.embed(["Banana"], EmbeddingModels.QA)
+    assert(len(e1.embeddings) == 1)
+    assert(len(e1.embeddings[0]) == 768)
+
+    e2 = nludb.embed(["This is a test"], EmbeddingModels.QA)
+    assert(len(e2.embeddings) == 1)
+    assert(len(e2.embeddings[0]) == 768)
+
+    e3 = nludb.embed(["This is a test"], EmbeddingModels.SIMILARITY)
+    e3b = nludb.embed(["Banana"], EmbeddingModels.SIMILARITY)
+    assert(len(e3.embeddings) == 1)
+    assert(len(e3.embeddings[0]) == 768)
+
+    e4 = nludb.embed(["This is a test"], EmbeddingModels.SIMILARITY)
+    assert(len(e4.embeddings) == 1)
+    assert(len(e4.embeddings[0]) == 768)
+
+    e5 = nludb.embed(["This is a test"], EmbeddingModels.PARAPHRASE)
+    e5b = nludb.embed(["Banana"], EmbeddingModels.PARAPHRASE)
+    assert(len(e5.embeddings) == 1)
+    assert(len(e5.embeddings[0]) == 768)
+
+    e6 = nludb.embed(["This is a test"], EmbeddingModels.PARAPHRASE)
+    assert(len(e6.embeddings) == 1)
+    assert(len(e6.embeddings[0]) == 768)
+
+    assert(e1.embeddings[0] == e2.embeddings[0])
+    assert(e3.embeddings[0] == e4.embeddings[0])
+    assert(e5.embeddings[0] == e6.embeddings[0])
+
+    assert(e1.embeddings[0] != e1b.embeddings[0])
+    assert(e3.embeddings[0] != e3b.embeddings[0])
+    assert(e5.embeddings[0] != e5b.embeddings[0])
+
+    assert(e1.embeddings[0] != e4.embeddings[0])
+    assert(e1.embeddings[0] != e6.embeddings[0])
+    assert(e3b.embeddings[0] != e5b.embeddings[0])
+    assert(e4.embeddings[0] != e6.embeddings[0])
+
+def test_basic_embedding_search():
+    nludb = _nludb()
+    docs = [
+        "Armadillo shells are bulletproof.",
+        "Dolphins sleep with one eye open.",
+        "Alfred Hitchcock was frightened of eggs.",
+        "Jonathan can help you with new employee onboarding",
+        "The code for the New York office is 1234",
+    ]
+    query = "Who should I talk to about new employee setup?"
+    results = nludb.embed_and_search(query, docs, EmbeddingModels.QA)
+    assert(len(results.hits) == 1)
+    assert(results.hits[0].value == "Jonathan can help you with new employee onboarding")
+
+
 def test_index_usage():
-    nludb = NLUDB(api_key = os.environ['NLUDB_KEY'])
+    nludb = _nludb()
     name = _random_name()
 
     index = nludb.create_index(
         name=name,
-        model=EmbeddingModels.DEFAULT_QA,
+        model=EmbeddingModels.QA,
         upsert=True
     )