+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+Version 0.1.4
+=============
+
+Initial release
+
+* FEATURE: Support for new semantic similarity model
+* DEMO: Semantic anomaly detection
+
 Version 0.1.3
 =============